Resuelvo problema de handlers
