+++ src/test/java/AlkemyWallet/AlkemyWallet/unit/AccountsUpdateTest.java
@@ -1,11 +1,10 @@
-package AlkemyWallet.AlkemyWallet;
+package AlkemyWallet.AlkemyWallet.unit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class LoginTest {
-
+public class AccountsUpdateTest {
     @Test
     void contextLoads() {
     }
