+++ src/test/java/AlkemyWallet/AlkemyWallet/unit/AccountsUsuarioTest.java
@@ -1,10 +1,10 @@
-package AlkemyWallet.AlkemyWallet;
+package AlkemyWallet.AlkemyWallet.unit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class FixedTermSimulateTest {
+public class AccountsUsuarioTest {
     @Test
     void contextLoads() {
     }
