+++ src/test/java/AlkemyWallet/AlkemyWallet/unit/AlkemyWalletApplicationTests.java
@@ -1,11 +1,13 @@
-package AlkemyWallet.AlkemyWallet;
+package AlkemyWallet.AlkemyWallet.unit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class TransactionDepositTest {
-    @Test
-    void contextLoads() {
-    }
+class AlkemyWalletApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
 }
