+++ src/test/java/AlkemyWallet/AlkemyWallet/unit/LoginTest.java
@@ -1,10 +1,11 @@
-package AlkemyWallet.AlkemyWallet;
+package AlkemyWallet.AlkemyWallet.unit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class AccountsUsuarioTest {
+public class LoginTest {
+
     @Test
     void contextLoads() {
     }
