+++ src/test/java/AlkemyWallet/AlkemyWallet/unit/RegisterTest.java
@@ -1,13 +1,11 @@
-package AlkemyWallet.AlkemyWallet;
+package AlkemyWallet.AlkemyWallet.unit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AlkemyWalletApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
+public class RegisterTest {
+    @Test
+    void contextLoads() {
+    }
 }
